+++ apps/web/src/app/[locale]/challenge/[slug]/@left/submissions/_components/index.tsx
