+++ apps/web/src/app/[locale]/challenge/[slug]/@left/submissions/_components/submissions-skeleton.tsx
