+++ apps/web/src/app/[locale]/challenge/[slug]/@left/submissions/loading.tsx
@@ -1,0 +1,1 @@
+export { SubmissionsSkeleton as default } from './_components/submissions-skeleton';
