+++ apps/web/src/app/[locale]/challenge/[slug]/@left/submissions/page.tsx
