+++ apps/web/src/app/[locale]/challenge/[slug]/@right/submissions/[submissionId]/page.tsx
@@ -1,0 +1,10 @@
+'use client';
+import { SubmissionOverview } from '../../../@left/submissions/_components/overview';
+
+export default function SubmissionPage({
+  params: { slug, submissionId },
+}: {
+  params: { slug: string; submissionId: string };
+}) {
+  return <SubmissionOverview submissionId={submissionId} />;
+}
