--- apps/web/src/app/[locale]/challenge/[slug]/submissions/loading.tsx
@@ -1,1 +1,0 @@
-export { SubmissionsSkeleton as default } from './[[...catchAll]]/_components/submissions-skeleton';
